Split custom_resources.go into multiple files (#413)
* split custom_resources.go into multiple files


* Remove redundant file


* remove unwanted comments
