+++ pkg/apis/openebs.io/v1alpha1/storage_pool.go
@@ -1,0 +1,58 @@
+/*
+Copyright 2018 The OpenEBS Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package v1alpha1
+
+import (
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+// +genclient
+// +genclient:noStatus
+// +genclient:nonNamespaced
+// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
+// +resource:path=storagepool
+
+// StoragePool describes a StoragePool.
+type StoragePool struct {
+	metav1.TypeMeta   `json:",inline"`
+	metav1.ObjectMeta `json:"metadata,omitempty"`
+
+	Spec StoragePoolSpec `json:"spec"`
+}
+
+// StoragePoolSpec is the spec for a StoragePool resource
+type StoragePoolSpec struct {
+	Name       string        `json:"name"`
+	Format     string        `json:"format"`
+	Mountpoint string        `json:"mountpoint"`
+	Nodename   string        `json:"nodename"`
+	Message    string        `json:"message"`
+	Path       string        `json:"path"`
+	Disks      DiskAttr      `json:"disks"`
+	PoolSpec   CStorPoolAttr `json:"poolSpec"`
+}
+
+// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
+// +resource:path=storagepools
+
+// StoragePoolList is a list of StoragePool resources
+type StoragePoolList struct {
+	metav1.TypeMeta `json:",inline"`
+	metav1.ListMeta `json:"metadata"`
+
+	Items []StoragePool `json:"items"`
+}
