+++ pkg/apis/openebs.io/v1alpha1/storage_pool_claim.go
@@ -1,0 +1,65 @@
+/*
+Copyright 2018 The OpenEBS Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package v1alpha1
+
+import (
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+// +genclient
+// +genclient:noStatus
+// +genclient:nonNamespaced
+// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
+// +resource:path=storagepoolclaim
+
+// StoragePoolClaim describes a StoragePoolClaim custom resource.
+type StoragePoolClaim struct {
+	metav1.TypeMeta   `json:",inline"`
+	metav1.ObjectMeta `json:"metadata,omitempty"`
+	Spec              StoragePoolClaimSpec   `json:"spec"`
+	Status            StoragePoolClaimStatus `json:"status"`
+}
+
+// StoragePoolClaimSpec is the spec for a StoragePoolClaimSpec resource
+type StoragePoolClaimSpec struct {
+	Name         string        `json:"name"`
+	Format       string        `json:"format"`
+	Mountpoint   string        `json:"mountpoint"`
+	Path         string        `json:"path"`
+	Type         string        `json:"type"`
+	NodeSelector []string      `json:"nodeSelector"`
+	Capacity     string        `json:"capacity"`
+	MaxPools     int16         `json:"maxPools"`
+	Disks        DiskAttr      `json:"disks"`
+	PoolSpec     CStorPoolAttr `json:"poolSpec"`
+}
+
+// StoragePoolClaim is for handling status of pool.
+type StoragePoolClaimStatus struct {
+	Phase string `json:"phase"`
+}
+
+// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
+// +resource:path=storagepoolclaims
+
+// StoragePoolClaimList is a list of StoragePoolClaim resources
+type StoragePoolClaimList struct {
+	metav1.TypeMeta `json:",inline"`
+	metav1.ListMeta `json:"metadata"`
+
+	Items []StoragePoolClaim `json:"items"`
+}
